requirements.txt
fastapi[standard]
fastapi-limiter
uvicorn[standard]
python-jose[cryptography]
passlib[bcrypt]
python-multipart
bcrypt
cryptography
sqlmodel
aiomysql
alembic
pymysql
